@@ -1,11 +1,11 @@
 <?php
 /**
- * NameValueList
+ * ComplianceSummaryInfo
  *
  * PHP version 7.2
  *
  * @category Class
- * @package  Ebay\Sell
+ * @package  Ebay\Sell\Compliance
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
  */
@@ -29,30 +29,30 @@
 namespace Ebay\Sell\Compliance\Model;
 
 use \ArrayAccess;
-use \Ebay\Sell\ObjectSerializer;
+use \Ebay\Sell\Compliance\ObjectSerializer;
 
 /**
- * NameValueList Class Doc Comment
+ * ComplianceSummaryInfo Class Doc Comment
  *
  * @category Class
- * @description This type is used to provide a name-value pair, including the identifying aspects of a product variation through the &lt;strong&gt;variationAspects&lt;/strong&gt; container.
- * @package  Ebay\Sell
+ * @description This type is used by each unique eBay marketplace and compliance type combination that is returned in the &lt;strong&gt;getListingViolationsSummary&lt;/strong&gt; response to indicate the total number of listing violations in regards to that eBay marketplace and compliance type.
+ * @package  Ebay\Sell\Compliance
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
  * @implements \ArrayAccess<TKey, TValue>
  * @template TKey int|null
  * @template TValue mixed|null
  */
-class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
+class ComplianceSummaryInfo implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
       *
       * @var string
       */
-    protected static $openAPIModelName = 'NameValueList';
+    protected static $openAPIModelName = 'ComplianceSummaryInfo';
 
     /**
       * Array of property to type mappings. Used for (de)serialization
@@ -60,8 +60,9 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
       * @var string[]
       */
     protected static $openAPITypes = [
-        'name' => 'string',
-        'value' => 'string'
+        'compliance_type' => 'string',
+        'marketplace_id' => 'string',
+        'listing_count' => 'int'
     ];
 
     /**
@@ -72,8 +73,9 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
       * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
-        'name' => null,
-        'value' => null
+        'compliance_type' => null,
+        'marketplace_id' => null,
+        'listing_count' => null
     ];
 
     /**
@@ -103,8 +105,9 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
      * @var string[]
      */
     protected static $attributeMap = [
-        'name' => 'name',
-        'value' => 'value'
+        'compliance_type' => 'complianceType',
+        'marketplace_id' => 'marketplaceId',
+        'listing_count' => 'listingCount'
     ];
 
     /**
@@ -113,8 +116,9 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
      * @var string[]
      */
     protected static $setters = [
-        'name' => 'setName',
-        'value' => 'setValue'
+        'compliance_type' => 'setComplianceType',
+        'marketplace_id' => 'setMarketplaceId',
+        'listing_count' => 'setListingCount'
     ];
 
     /**
@@ -123,8 +127,9 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
      * @var string[]
      */
     protected static $getters = [
-        'name' => 'getName',
-        'value' => 'getValue'
+        'compliance_type' => 'getComplianceType',
+        'marketplace_id' => 'getMarketplaceId',
+        'listing_count' => 'getListingCount'
     ];
 
     /**
@@ -184,8 +189,9 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
      */
     public function __construct(array $data = null)
     {
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['value'] = isset($data['value']) ? $data['value'] : null;
+        $this->container['compliance_type'] = $data['compliance_type'] ?? null;
+        $this->container['marketplace_id'] = $data['marketplace_id'] ?? null;
+        $this->container['listing_count'] = $data['listing_count'] ?? null;
     }
 
     /**
@@ -213,49 +219,73 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
 
 
     /**
-     * Gets name
+     * Gets compliance_type
      *
      * @return string|null
      */
-    public function getName()
+    public function getComplianceType()
     {
-        return $this->container['name'];
+        return $this->container['compliance_type'];
     }
 
     /**
-     * Sets name
+     * Sets compliance_type
      *
-     * @param string|null $name This is the name of the variation aspect, or the name of the category of information that is returned through the name-value pair. The type of information that appears here will vary based on the compliance type and type of violation.
+     * @param string|null $compliance_type This enumeration value indicates the type of compliance. See ComplianceTypeEnum for more information on each compliance type. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/sell/compliance/types/com:ComplianceTypeEnum'>eBay API documentation</a>
      *
      * @return self
      */
-    public function setName($name)
+    public function setComplianceType($compliance_type)
     {
-        $this->container['name'] = $name;
+        $this->container['compliance_type'] = $compliance_type;
 
         return $this;
     }
 
     /**
-     * Gets value
+     * Gets marketplace_id
      *
      * @return string|null
      */
-    public function getValue()
+    public function getMarketplaceId()
     {
-        return $this->container['value'];
+        return $this->container['marketplace_id'];
     }
 
     /**
-     * Sets value
+     * Sets marketplace_id
      *
-     * @param string|null $value This is the value of the variation aspect (in name field), or the value of the category of information that is returned through the name-value pair. The type of information that appears here will vary based on the compliance type and type of violation.
+     * @param string|null $marketplace_id This enumeration value indicates the eBay marketplace where the listing violations exist. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/sell/compliance/types/bas:MarketplaceIdEnum'>eBay API documentation</a>
      *
      * @return self
      */
-    public function setValue($value)
+    public function setMarketplaceId($marketplace_id)
     {
-        $this->container['value'] = $value;
+        $this->container['marketplace_id'] = $marketplace_id;
+
+        return $this;
+    }
+
+    /**
+     * Gets listing_count
+     *
+     * @return int|null
+     */
+    public function getListingCount()
+    {
+        return $this->container['listing_count'];
+    }
+
+    /**
+     * Sets listing_count
+     *
+     * @param int|null $listing_count This integer value indicates the number of eBay listings that are currently violating the compliance type indicated in the complianceType field, for the eBay marketplace indicated in the marketplaceId field.
+     *
+     * @return self
+     */
+    public function setListingCount($listing_count)
+    {
+        $this->container['listing_count'] = $listing_count;
 
         return $this;
     }
@@ -280,7 +310,7 @@ class NameValueList implements ModelInterface, ArrayAccess, \JsonSerializable
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
